# src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Heimdall API",
    "description": "Multi-tenant authentication and authorization service with password and WebAuthn/passkey support.",
    "version": "1.0.0",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base path"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from login or refresh"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Project API key"
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "description": "1-based page number.",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Results per page. Values above the maximum are clamped.",
        "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        },
        "required": ["message"]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "username": { "type": "string" },
          "role": { "type": "string", "enum": ["owner", "admin", "manager", "member"] },
          "membershipId": { "type": "string" },
          "joinedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Membership": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "userId": { "type": "object" },
          "projectId": { "type": "string" },
          "role": { "type": "string", "enum": ["owner", "admin", "manager", "member"] },
          "status": { "type": "string", "enum": ["active", "pending", "suspended"] },
          "metadata": { "type": "object" },
          "invitedBy": { "type": "string" },
          "joinedAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List all users in project",
        "description": "Returns a page of active users and their membership info for the current project.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "Paged user list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/User" }
                    },
                    "page": { "type": "integer", "example": 1 },
                    "limit": { "type": "integer", "example": 50 },
                    "total": {
                      "type": "integer",
                      "description": "Total matching records across all pages.",
                      "example": 137
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
    "/memberships": {
      "get": {
        "tags": ["Memberships"],
        "summary": "List project members",
        "description": "Returns a page of active and pending memberships for the current project.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "Paged membership list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "members": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Membership" }
                    },
                    "page": { "type": "integer", "example": 1 },
                    "limit": { "type": "integer", "example": 50 },
                    "total": {
                      "type": "integer",
                      "description": "Total matching records across all pages.",
                      "example": 137
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
  }
}